+++ adoption/src/main/java/com/refactoringlife/adoption/features/AdoptionActivity.kt
@@ -3,8 +3,8 @@ package com.refactoringlife.adoption.features
 import android.os.Bundle
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
-import com.refactoringlife.adoption.features.welcomeOnboarding.presentation.screen.WelcomeOnboardingScreen
-import com.refactoringlife.adoption.features.welcomeOnboarding.presentation.utils.WelcomeOnboardingOption
+import com.refactoringlife.adoption.features.welcome.presentation.screen.WelcomeOnboardingScreen
+import com.refactoringlife.adoption.utils.WelcomeOptions
 
 class AdoptionActivity : ComponentActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
@@ -12,7 +12,7 @@ class AdoptionActivity : ComponentActivity() {
 
         setContent {
             WelcomeOnboardingScreen(
-                selectedOption = WelcomeOnboardingOption.GALLERY,
+                selectedOption = WelcomeOptions.GALLERY,
                 onOptionSelected = {},
                 onAccept = {},
                 onSkip = {}
