--- adoption/src/main/java/com/refactoringlife/adoption/features/welcomeOnboarding/presentation/composables/WelcomeOnboardingRadioButton.kt
@@ -1,44 +1,0 @@
-package com.refactoringlife.adoption.features.welcomeOnboarding.presentation.composables
-
-import androidx.compose.foundation.clickable
-import androidx.compose.foundation.layout.Row
-import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.padding
-import androidx.compose.material3.RadioButton
-import androidx.compose.material3.RadioButtonDefaults
-import androidx.compose.material3.Text
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.Alignment
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.unit.dp
-import androidx.compose.ui.unit.sp
-import com.refactoringlife.core.presentation.theme.HuellaPurple
-
-@Composable
-fun WelcomeOnboardingRadioButton(
-    text: String,
-    selected: Boolean,
-    onClick: () -> Unit
-) {
-    Row(
-        modifier = Modifier
-            .fillMaxWidth()
-            .padding(vertical = 8.dp)
-            .clickable { onClick() },
-        verticalAlignment = Alignment.CenterVertically
-    ) {
-        RadioButton(
-            selected = selected,
-            onClick = onClick,
-            colors = RadioButtonDefaults.colors(
-                selectedColor = HuellaPurple
-            )
-        )
-        Text(
-            text = text,
-            color = Color.Black,
-            fontSize = 18.sp
-        )
-    }
-}
