--- adoption/src/main/java/com/refactoringlife/adoption/features/welcomeOnboarding/presentation/utils/WelcomeOnboardingOption.kt
@@ -1,7 +1,0 @@
-package com.refactoringlife.adoption.features.welcomeOnboarding.presentation.utils
-
-enum class WelcomeOnboardingOption {
-    GALLERY,
-    FOUNDATIONS,
-    REPORT
-}
