+++ adoption/src/main/java/com/refactoringlife/adoption/utils/WelcomeOptions.kt
@@ -1,0 +1,7 @@
+package com.refactoringlife.adoption.utils
+
+enum class WelcomeOptions {
+    GALLERY,
+    FOUNDATIONS,
+    REPORT
+}
