+++ auth/src/main/java/com/refactoringlife/auth/features/activity/presenter/viewModel/AuthViewModel.kt
@@ -1,11 +1,11 @@
-package com.refactoringlife.auth.features.onboarding.presentation.viewmodel
+package com.refactoringlife.auth.features.activity.presenter.viewModel
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import com.refactoringlife.auth.features.onboarding.domain.blocs.OnboardingBaseBloc
-import com.refactoringlife.auth.features.onboarding.domain.blocs.OnboardingEvent
-import com.refactoringlife.auth.features.onboarding.domain.blocs.OnboardingBlocs
-import com.refactoringlife.auth.features.onboarding.domain.state.OnboardingState
+import com.refactoringlife.auth.features.activity.domain.blocs.AuthBaseBloc
+import com.refactoringlife.auth.features.activity.domain.blocs.AuthBlocs
+import com.refactoringlife.auth.features.activity.domain.blocs.AuthEvent
+import com.refactoringlife.auth.features.activity.domain.state.AuthState
 import com.refactoringlife.core.data.datastore.AppPreferencesRepository
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.Dispatchers
@@ -16,23 +16,23 @@ import kotlinx.coroutines.withContext
 import javax.inject.Inject
 
 @HiltViewModel
-class OnboardingViewModel @Inject constructor(
+class AuthViewModel @Inject constructor(
     appPreferencesRepository: AppPreferencesRepository
 ) : ViewModel() {
 
-    private val blocs: List<OnboardingBaseBloc> = OnboardingBlocs.getOnboardingBlocs(appPreferencesRepository)
+    private val blocs: List<AuthBaseBloc> = AuthBlocs.Companion.getAuthBlocs(appPreferencesRepository)
 
-    private val _state = MutableStateFlow(OnboardingState())
-    val state: StateFlow<OnboardingState> = _state
+    private val _state = MutableStateFlow(AuthState())
+    val state: StateFlow<AuthState> = _state
 
-    fun sendEvent(event: OnboardingEvent) {
+    fun sendEvent(event: AuthEvent) {
         viewModelScope.launch {
             val bloc = blocs.firstOrNull { it.canHandle(event) } ?: return@launch
             bloc.handle(event) { reducer -> updateState(reducer) }
         }
     }
 
-    private suspend fun updateState(reducer: suspend (OnboardingState) -> OnboardingState) {
+    private suspend fun updateState(reducer: suspend (AuthState) -> AuthState) {
         withContext(Dispatchers.Main) {
             val next = reducer(_state.value)
             _state.value = next
