+++ auth/src/main/java/com/refactoringlife/auth/features/onboarding/domain/blocs/OnboardingBaseBloc.kt
@@ -1,0 +1,11 @@
+package com.refactoringlife.auth.features.onboarding.domain.blocs
+
+import com.refactoringlife.auth.features.onboarding.domain.state.OnboardingState
+
+typealias OnboardingStateUpdater = suspend (suspend (OnboardingState) -> OnboardingState) -> Unit
+
+interface OnboardingBaseBloc {
+    fun canHandle(event: OnboardingEvent): Boolean
+    suspend fun handle(event: OnboardingEvent, update: OnboardingStateUpdater)
+}
+
