+++ auth/src/main/java/com/refactoringlife/auth/features/onboarding/domain/blocs/OnboardingEvent.kt
@@ -1,0 +1,6 @@
+package com.refactoringlife.auth.features.onboarding.domain.blocs
+
+sealed class OnboardingEvent {
+    object CompleteOnboarding : OnboardingEvent()
+}
+
