+++ auth/src/main/java/com/refactoringlife/auth/features/onboarding/domain/state/OnboardingState.kt
@@ -1,0 +1,9 @@
+package com.refactoringlife.auth.features.onboarding.domain.state
+
+data class OnboardingState(
+    val completed: Boolean = false,
+    val loading: Boolean = false,
+    val error: Boolean = false,
+    val errorMessage: String? = null
+)
+
