+++ auth/src/main/java/com/refactoringlife/auth/features/resetPassword/domain/bloc/ResetPasswordBaseBloc.kt
@@ -1,0 +1,10 @@
+package com.refactoringlife.auth.features.resetPassword.domain.bloc
+
+import com.refactoringlife.auth.features.resetPassword.domain.state.ResetPasswordState
+
+typealias ResetPasswordStateUpdater = suspend (suspend (ResetPasswordState) -> ResetPasswordState) -> Unit
+
+interface ResetPasswordBaseBloc {
+    fun canHandle(event: ResetPasswordEvent): Boolean
+    suspend fun handle(event: ResetPasswordEvent, update: ResetPasswordStateUpdater)
+}
