+++ auth/src/main/java/com/refactoringlife/auth/features/resetPassword/domain/bloc/ResetPasswordBlocs.kt
@@ -1,0 +1,7 @@
+package com.refactoringlife.auth.features.resetPassword.domain.bloc
+
+class ResetPasswordBlocs{
+    companion object {
+        fun getResetPasswordBlocs() = listOf(HandleResetPasswordBloc())
+    }
+}
