+++ auth/src/main/java/com/refactoringlife/auth/features/resetPassword/domain/state/ResetPasswordState.kt
@@ -1,0 +1,7 @@
+package com.refactoringlife.auth.features.resetPassword.domain.state
+
+data class ResetPasswordState(
+    val showModal : Boolean = false,
+    val password : String = "",
+    val confirmPassword : String = ""
+)
