--- auth/src/main/java/com/refactoringlife/auth/features/resetpassword/domain/bloc/ResetPasswordBlocs.kt
@@ -1,7 +1,0 @@
-package com.refactoringlife.auth.features.resetpassword.domain.bloc
-
-class ResetPasswordBlocs{
-    companion object {
-        fun getResetPasswordBlocs() = listOf(HandleResetPasswordBloc())
-    }
-}
