--- auth/src/main/java/com/refactoringlife/auth/features/resetpassword/domain/bloc/ResetPasswordEvent.kt
@@ -1,9 +1,0 @@
-package com.refactoringlife.auth.features.resetpassword.domain.bloc
-
-sealed class ResetPasswordEvent {
-    data class ResetPassword(
-        val password : String = "",
-        val confirmPassword : String = "",
-        val token : String = ""
-    ) : ResetPasswordEvent()
-}
