--- auth/src/main/java/com/refactoringlife/auth/features/resetpassword/domain/state/ResetPasswordState.kt
@@ -1,7 +1,0 @@
-package com.refactoringlife.auth.features.resetpassword.domain.state
-
-data class ResetPasswordState(
-    val showModal : Boolean = false,
-    val password : String = "",
-    val confirmPassword : String = ""
-)
