+++ core/src/main/java/com/refactoringlife/core/common/utils/Constants.kt
@@ -3,4 +3,5 @@ package com.refactoringlife.core.common.utils
 object Constants {
     const val LOADING_TAG = "LOADING_TAG"
     const val EMPTY = ""
+    const val ID = "id"
 }
